Move files to different packages under example/geonames
