+++ examples/geonames/gradle.properties
@@ -1,0 +1,10 @@
+mlHost=oscar
+mlAppName=marklogic-spring-batch-test
+mlJobRepositoryName=marklogic-spring-batch-test
+mlRestPort=8200
+mlUsername=admin
+mlPassword=admin
+mlManageUsername=admin
+mlManagePassword=admin
+mlConfigDir=src/test/ml-config
+mlContentForestsPerHost=1
