+++ examples/geonames/src/main/java/com/marklogic/example/geonames/CountryFieldSetMapper.java
@@ -1,4 +1,4 @@
-package com.marklogic.spring.batch.geonames;
+package com.marklogic.example.geonames;
 
 import org.springframework.batch.item.file.mapping.FieldSetMapper;
 import org.springframework.batch.item.file.transform.FieldSet;
