+++ examples/geonames/src/main/java/com/marklogic/example/geonames/GeonameFieldSetMapper.java
@@ -1,4 +1,4 @@
-package com.marklogic.spring.batch.geonames;
+package com.marklogic.example.geonames;
 
 import java.util.List;
 import java.util.ArrayList;
