+++ examples/geonames/src/main/java/org/geonames/Country.java
