+++ examples/geonames/src/main/java/org/geonames/Geoname.java
