+++ examples/geonames/src/test/java/org/example/IngestGeonamesToMarkLogicTest.java
@@ -1,8 +1,11 @@
 package org.example;
 
+import com.marklogic.client.spring.BasicConfig;
 import com.marklogic.spring.batch.test.AbstractJobTest;
 import org.junit.Test;
+import org.springframework.test.context.ContextConfiguration;
 
+@ContextConfiguration(classes = {BasicConfig.class})
 public class IngestGeonamesToMarkLogicTest extends AbstractJobTest {
 
     @Test
